% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/song.R
\name{beer_song}
\alias{beer_song}
\title{Write a song.}
\usage{
beer_song(n = 99, vessel = "bottles", liquid = "beer",
  surface = "wall")
}
\arguments{
\item{n}{number of vessel.}

\item{vessel}{vessel}
}
\value{
a song
}
\description{
Write a song.
}
\examples{
beer_song(n = 5, vessel = "bottles", liquid = "beer", surface = "wall")

}
